Implement connection upgrade
Moving WebSocket server creation from `server.js` to `jobs.js` because that's where it's actually needed.

`handleUpgrades` must be set to `true` when creating the restify server, otherwise incoming requests with `Connection: Upgrade` header are not passed down the regular routing chain. See http://restify.com/docs/home/#upgrade-requests

Specifying the HTTP server when creating the WebSocket server results in accepting ANY connection to `ws://localhost`. Instead, we start the WebSocket server in `noServer` mode and then manually claim and upgrade an appropriate incoming connection.

Because I haven't found a way to initiate the WebSocket upgrade with a POST in the browsers' WebSocket API, I switched the endpoint to a GET (at least for now).